This is a large check-in which converts our build system to use autoconf.
See the file INSTALL.email for details on the build process.

Changes to source code are essentially limited to macro conventions
changes: I have converted to using macros like ENABLE_RMS which
corresponds to the configure argument

	configure --enable-rms

and so on.


Tested fully on lambda.  Not tested elsewhere as much as I would like
(due to Q "upgrade" etc.) but I need to get this in now!