Fix-up for non-primitive basic types.  MPI_Get_elements() should return the number of "primitive" types, and apparently we are supposed to say that MPI_2INT & co. consist of two primitive types.
This is a kludge.  Doing something better is tricky because we would
ideally use MPI_Type_struct to create these types during
initialization, but can't because other aspects of initialization are
not yet complete.

My preferred solution would be to separate the type creation logic
into two parts: ULM type creation, and layered on top the MPI specific
book-keeping.  This is a fairly major task however ... hence this
nastiness.