Propagating bug fix from patch branch.
On crescendo (CSS-3's linux/ia32/quadrics cluster) using gcc 2.96, our
fortran binding scheme was causing the symbols for the default C
attribute functions (mpi_null_delete_fn_c etc.) to be redefined as
weak bindings to the fortran functions.

This was causing a large number of communicator tests to fail with SIGSEGV.

This is not a problem elsewhere -- it depends on the precendence of
#define and #pragma used by the compiler/preprocessor.

In particular Q and lambda (with gcc 3.0) are OK, so we don't have to
rush the fix into the 1.3.1 releases on these platforms.