elan hw bcast shared memory optimziation.
With shared memory enabled, It works in many cases, but still hangs on
a few of the intel bcast tests.  Weikuan has agreed to track down
those problems in the near future.

To turn on original code:

add USE_ELAN_COLL to CPPFLAGS, and use the -qf hw option.

To turn on shared memory:  add USE_ELAN_SHARED, which just effects
3 small blocks of code.