- added useful assert() for debugging problems with setting messageDone to true multiple times...which is very bad for multithreaded code...
- significant rework of IB acknowledgement code that should probably be pushed
into other paths (verified under 192 process runs of multithreaded MPQC code)

The acknowledgement code has the following major changes:

1) If a fragment's sequence number is greater than the peer's reported largest
in-order received sequence number, then it will be scheduled for retransmission
if a) the current time is greater than the calculated time to resend; b)
at least weighted_average_rtt seconds have passed since the last transmission;
c) the receive completion queue is empty indicating that there are no
possible acks that have not been received (UD only). If the fragment is not
retransmitted, then the calculated time to resend is used to update the
message's earliest time to resend.

weighted_average_rtt is calculated for every destination peer, and stored
in the sender ackinfo reliabilityInfo structure. It is a weighted average
of (time[ack] - time[sent]).

2) If a fragment has not been retransmitted in ib_state.nack_lost_retrans_time
seconds, then the fragment is unconditionally scheduled for retransmission.
This overcomes the ack hole that's existed for some time...

These changes have reduced extra retransmission and acknowledgement by a
factor of 3 to 5 in MPQC runs. - Mitch