Problem encountered in QCD burn-in code with an MPI_Irecv from MPI_PROC_NULL, followed by an MPI_Test.
Our logic in MPI_Test for this case was wrong -- we check for a
valid request first, then for the placeholder proc_null request.  We
should do it the other way around as we do for MPI_Wait.

Moved the check for this case into ulm_check.cc and also cleaned up
MPI_Test and MPI_Wait to be more symmetrical.