TotalView symbol mapping problem:
Moving TotalView debug gate kludge from mpi.h to a separate file
debug_gate.c.  This is compiled as a separate target and installed to

       $(MPI_ROOT)/lib/debug_gate.o

Users of the latest TotalView on lightning (and similar systems??)
should link this object file with their application to ensure that
TotalView can find the MPIR_debug_gate symbol.